third_party/blink/web_tests/virtual/omt-worker-fetch/http/tests/security/cors-rfc1918/addressspace-sharedworker-basic-expected.txt

This is a testharness.js-based test.
FAIL http://localhost:8000 = 'local' assert_equals: origin expected (string) "http://localhost:8000" but got (undefined) undefined
FAIL http://127.0.0.1:8000 = 'local' assert_equals: origin expected (string) "http://127.0.0.1:8000" but got (undefined) undefined
PASS http://example.test:8000 = 'private'
Harness: the test ran to completion.
